seperate the node struct to a file
