+++ core/src/repair/repair_service.rs
@@ -1262,7 +1262,7 @@ mod test {
     use {
         super::*,
         crate::repair::quic_endpoint::RemoteRequest,
-        solana_gossip::{cluster_info::Node, contact_info::ContactInfo},
+        solana_gossip::{node::Node, contact_info::ContactInfo},
         solana_keypair::Keypair,
         solana_ledger::{
             blockstore::{
